fix(minimax_m3): PR #110 review batch -- NaN-poisoning masks, MXFP8 kernel perf, parser robustness
Correctness (both reviews reproduced these on live hardware):

- [HIGH] Restore the reference pos-masks on all four sparse-attend K/V loads:
  the pool slab is torch.empty and the forced local block reads the partial
  newest page EVERY decode step, so unmasked tail rows (recycled-allocator
  NaN/Inf bit patterns) NaN'd the whole output row. BSAKVCache now also
  zero-fills the K/V slab on (re)alloc as defense-in-depth, the index-score
  kernels' zero-init reliance is documented, and the kernel tests fill
  unwritten rows with NaN (zero-filled slabs are exactly why the bug survived
  them) + parametrize the checkpoint's topk=16.
- [HIGH] MiniMaxM3Detector rewrite: wire order preserved (same-chunk trailing
  text defers past the call), a closed wrapper loops back through the idle scan
  (second block / split marker in the residue held, never leaked), the buffer
  keeps the wrapper opener in-block so finish_streaming suppresses truncated
  markup and recover_truncated_call can salvage the call, lenient element
  scanning (stray text skipped instead of nuking all parsed params, empty
  element -> "" not {}, repeated siblings -> list, truncated trailing element
  salvages complete siblings), and detect_and_parse handles multiple wrappers
  with inter-block text.
- [HIGH] MiniMaxM3ReasoningParser: strip the single leading bare </mm:think>
  an adaptive non-thinking turn starts with (position 0 only, one-shot and
  streaming with prefix hold) -- it leaked verbatim into content on the
  default gear's most common path.
- [HIGH] _pick_inner_backend resolves through the engine's own FULL auto-tree
  (a private copy gated "fa,fi" on the open-ended is_sm90_supported and
  crashed sm_12x + sgl_kernel boxes at construction), logs the resolved
  backend, and validates the env override against the capability matrix.
- [MED] GemmaPlusOneRMSNorm collapses per-head 3D views to 2D before the
  flashinfer call (its CUDA binding is CHECK_DIM(2)).
- [MED] Dummy-weight uint8 fill scoped by name to weight_scale_inv (the blanket
  127 fill turned GGUF qweight-embedded fp16 scales into NaN).
- [MED] BSA index slab keys off the engine dtype (fp16 serving crashed on the
  index_put_ dtype mismatch); rebuild nulls the K/V slab too when the index
  realloc fails; router gate stored fp32 (top-4 tie-break boundary); env
  serving-mode switches (SPARSE=0 / MAX_LAYERS / MXFP8) log once at parse time
  (covers the FTW serve path); parse-time asserts for the geometry the kernels
  hardcode (block 128, index_dim == head_dim, topk < 64, local_blocks >= 1,
  swiglu_beta == 1, sigmoid scoring).

Performance:

- [HIGH] MXFP8 was the default AND slower than the bf16 ablation (decode GEMV
  424-768 GB/s from a per-element scale gather + exp2; fused GEMM 30x off
  cuBLAS). The GEMV now loads the e8m0 codes once per (BLOCK_N, BLOCK_K/32)
  tile and broadcasts in registers -- measured 2-4x FASTER than bf16 cuBLAS on
  the M3 decode shapes (18.5us vs 75.7us for qkv) -- and M>1 dequantizes to a
  bf16 transient + cuBLAS (~7.5x faster than the deleted fused GEMM). The
  mxfp8 default now earns its keep instead of taxing it.
- [MED] fi workspace derived from flashinfer's actual split bound (TP-local
  heads x padded_batch(2*SM/kv_local) x cta_tile_q x head_dim), floored at the
  old 256 MiB -- the token-count heuristic over-allocated ~2x on wide-head
  models (GLM-4.7: 512 MiB vs ~198 needed) out of the post-KV headroom.
- [MED] Decode staging gathers through the every-128th-column page-table VIEW
  (the full-width materialization was 2 MiB/req/step at long contexts);
  prefill's per-request addressing plan is built once per forward instead of
  per sparse layer (was ~171 stream-serializing H2D copies at bs=1); the
  decode top-k split is computed in-kernel from the row's LIVE block count
  (the staged-width split serialized short contexts into chunk 0).

Also rebased onto main (picks up the _page_table_width fix that P=128
geometries want). Verified: 1096 tests green; dummy smokes (sparse + dense
ablation); real-checkpoint hybrid serving on the RTX PRO 6000 -- AIME 4/4,
25.5 tok/s single-stream, multi-invoke tool calls through the rewritten
streaming parser.

+++ python/freetoken/models/minimax_m3/weight.py
@@ -203,7 +203,11 @@ def iter_weights(
 
             if layer in args.moe_layer_ids:
                 m_src, m_dst = f"{src}.block_sparse_moe", f"{dst}.block_sparse_moe"
-                yield f"{m_dst}.gate.weight", reader.get(f"{m_src}.gate.weight")
+                # fp32 like the bias: both sit on the top-4 selection boundary.
+                yield (
+                    f"{m_dst}.gate.weight",
+                    reader.get(f"{m_src}.gate.weight").to(torch.float32),
+                )
                 # fp32 in the checkpoint AND the module (top-k selection boundary).
                 yield (
                     f"{m_dst}.e_score_correction_bias",
